refactor: streamline map component logic and introduce custom hooks for data handling
