+++ src/components/Map/map-screen.web.tsx
@@ -1,7 +1,5 @@
-import { trackEvent } from '@aptabase/react-native'
 import type BottomSheet from '@gorhom/bottom-sheet'
 import type { BottomSheetModal } from '@gorhom/bottom-sheet'
-import { useQuery } from '@tanstack/react-query'
 import {
 	Layer,
 	// biome-ignore lint/suspicious/noShadowRestrictedNames: TODO
@@ -10,9 +8,8 @@ import {
 	NavigationControl,
 	Source
 } from '@vis.gl/react-maplibre'
-import { toast } from 'burnt'
-import { router, useLocalSearchParams, useNavigation } from 'expo-router'
-import type { Feature, FeatureCollection, Position } from 'geojson'
+import { router, useNavigation } from 'expo-router'
+import type { Position } from 'geojson'
 import maplibregl from 'maplibre-gl'
 import type React from 'react'
 import { use, useCallback, useEffect, useMemo, useRef, useState } from 'react'
@@ -36,12 +33,6 @@ import {
 	UnistylesRuntime,
 	useStyles
 } from 'react-native-unistyles'
-import API from '@/api/authenticated-api'
-import NeulandAPI from '@/api/neuland-api'
-import {
-	NoSessionError,
-	UnavailableSessionError
-} from '@/api/thi-session-handler'
 import { UserKindContext } from '@/components/contexts'
 import ErrorView from '@/components/Error/error-view'
 import { BottomSheetDetailModal } from '@/components/Map/bottom-sheet-detail-modal'
@@ -49,38 +40,19 @@ import MapBottomSheet from '@/components/Map/bottom-sheet-map'
 import FloorPicker from '@/components/Map/floor-picker'
 import { MapContext } from '@/contexts/map'
 import { USER_GUEST } from '@/data/constants'
+import { useMapGeoJsonFilters } from '@/hooks/useMapGeoJsonFilters'
+import { useMapQueries } from '@/hooks/useMapQueries'
+import { useMapRoomSelection } from '@/hooks/useMapRoomSelection'
 import useRouteParamsStore from '@/hooks/useRouteParamsStore'
-import { type FeatureProperties, Gebaeude } from '@/types/asset-api'
 import {
 	type ClickedMapElement,
 	type RoomData,
 	SEARCH_TYPES
 } from '@/types/map'
 import type { NormalizedLecturer } from '@/types/utils'
-import { formatISODate, formatISOTime } from '@/utils/date-utils'
-import { normalizeLecturers } from '@/utils/lecturers-utils'
-import {
-	filterAvailableRooms,
-	filterEtage,
-	getBuildingData,
-	getOngoingOrNextEvent,
-	getRoomData
-} from '@/utils/map-screen-utils'
-import {
-	BUILDINGS,
-	FLOOR_ORDER,
-	FLOOR_SUBSTITUTES,
-	filterRooms,
-	getCenter,
-	getCenterSingle,
-	getIcon,
-	getRoomOpenings,
-	INGOLSTADT_CENTER,
-	NEUBURG_CENTER
-} from '@/utils/map-utils'
-import { loadTimetable } from '@/utils/timetable-utils'
-import { LoadingState, roomNotFoundToast } from '@/utils/ui-utils'
-import packageInfo from '../../../package.json'
+import { getBuildingData, getRoomData } from '@/utils/map-screen-utils'
+import { INGOLSTADT_CENTER, NEUBURG_CENTER } from '@/utils/map-utils'
+import { LoadingState } from '@/utils/ui-utils'
 import LoadingIndicator from '../Universal/loading-indicator'
 import { modalSection } from './modal-sections'
 import 'maplibre-gl/dist/maplibre-gl.css'
@@ -94,7 +66,6 @@ const MapScreen = (): React.JSX.Element => {
 	const [mapLoadState, setMapLoadState] = useState(LoadingState.LOADING)
 	const { styles, theme } = useStyles(stylesheet)
 	const isDark = UnistylesRuntime.themeName === 'dark'
-	const params = useLocalSearchParams<{ room: string }>()
 	const { userKind, userFaculty } = use(UserKindContext)
 	const [mapCenter, setMapCenter] = useState(INGOLSTADT_CENTER)
 	const { t, i18n } = useTranslation('common')
@@ -103,21 +74,16 @@ const MapScreen = (): React.JSX.Element => {
 	const currentPosition = useSharedValue(0)
 	const currentPositionModal = useSharedValue(0)
 	const {
-		localSearch,
 		clickedElement,
 		setClickedElement,
 		availableRooms,
-		setAvailableRooms,
 		roomOpenings,
-		setRoomOpenings,
 		currentFloor,
-		setCurrentFloor,
-		setNextLecture
+		setCurrentFloor
 	} = use(MapContext)
 	const [_disableFollowUser, setDisableFollowUser] = useState(false)
 	const [showAllFloors, setShowAllFloors] = useState(false)
 	const mapRef = useRef(null)
-	const currentDate = new Date()
 
 	enum Locations {
 		IN = 'Ingolstadt',
@@ -160,55 +126,28 @@ const MapScreen = (): React.JSX.Element => {
 		bottomSheetModalRef.current?.present()
 	}, [])
 
-	const { data: mapOverlay, error: overlayError } = useQuery<FeatureCollection>(
-		{
-			queryKey: ['mapOverlay', packageInfo.version],
-			queryFn: async () => await NeulandAPI.getMapOverlay(),
-			staleTime: 1000 * 60 * 60 * 24 * 7, // 1 week
-			gcTime: 1000 * 60 * 60 * 24 * 60, // 60 days
-			networkMode: 'always'
-		}
-	)
+	const { mapOverlay, overlayError, lecturers, allRooms, buildingGeoJSON } =
+		useMapQueries()
 
-	useEffect(() => {
-		if (overlayError != null) {
-			toast({
-				title: t('toast.mapOverlay', { ns: 'common' }),
-				preset: 'error',
-				duration: 3,
-				from: 'top'
-			})
-		}
-	}, [overlayError])
-
-	const { data: timetable } = useQuery({
-		queryKey: ['timetableV2', userKind],
-		queryFn: loadTimetable,
-		staleTime: 1000 * 60 * 10, // 10 minutes
-		gcTime: 1000 * 60 * 60 * 24 * 7, // 1 week
-		retry(failureCount, error) {
-			const ignoreErrors = [
-				'"Time table does not exist" (-202)',
-				'Timetable is empty'
-			]
-			if (ignoreErrors.includes(error.message)) {
-				return false
-			}
-			return failureCount < 2
-		},
-		enabled: userKind !== USER_GUEST
+	const { selectRoom } = useMapRoomSelection({
+		allRooms,
+		mapLoadState,
+		handlePresentModalPress,
+		bottomSheetRef,
+		notificationColor: theme.colors.notification
 	})
 
-	const { data: lecturers } = useQuery({
-		queryKey: ['allLecturers'],
-		queryFn: async () => {
-			const rawData = await API.getLecturers('0', 'z')
-			const data = normalizeLecturers(rawData)
-			return data
-		},
-		staleTime: 1000 * 60 * 30, // 30 minutes
-		gcTime: 1000 * 60 * 60 * 24 * 7, // 7 days
-		enabled: userKind !== USER_GUEST
+	const {
+		uniqueEtages,
+		filteredGeoJSON,
+		availableFilteredGeoJSON,
+		hasFilteredRooms,
+		hasAvailableFilteredRooms
+	} = useMapGeoJsonFilters({
+		mapOverlay,
+		allRooms,
+		currentFloor,
+		availableRooms
 	})
 
 	useEffect(() => {
@@ -220,107 +159,6 @@ const MapScreen = (): React.JSX.Element => {
 			subscription.remove()
 		}
 	})
-
-	useEffect(() => {
-		if (timetable == null) {
-			setNextLecture([])
-			return
-		}
-		const ongoingOrNextEvent = getOngoingOrNextEvent(timetable)
-		if (ongoingOrNextEvent.length > 0) {
-			setNextLecture(ongoingOrNextEvent)
-		}
-	}, [timetable, userKind])
-
-	const allRooms: FeatureCollection = useMemo(() => {
-		if (mapOverlay == null) {
-			console.debug('No map overlay data')
-			return { type: 'FeatureCollection', features: [] }
-		}
-
-		const rooms = mapOverlay.features.flatMap((feature) => {
-			const { type, id, geometry, properties } = feature
-
-			if (
-				geometry == null ||
-				properties == null ||
-				geometry.type !== 'Polygon' ||
-				!('coordinates' in geometry) ||
-				geometry.coordinates == null ||
-				geometry.coordinates.length === 0
-			) {
-				return []
-			}
-
-			if (properties.Ebene in FLOOR_SUBSTITUTES) {
-				properties.Ebene = FLOOR_SUBSTITUTES[properties.Ebene as string]
-			}
-			if (!FLOOR_ORDER.includes(properties.Ebene as string)) {
-				FLOOR_ORDER.push(properties.Ebene as string)
-			}
-			return {
-				type,
-				id,
-				properties: {
-					...properties,
-					rtype: SEARCH_TYPES.ROOM,
-					center: getCenterSingle(geometry.coordinates),
-					icon: getIcon(SEARCH_TYPES.ROOM, {
-						result: { item: { properties } }
-					})
-				} as unknown as FeatureProperties,
-				geometry
-			}
-		})
-		const buildings = BUILDINGS.map((building) => {
-			const buildingRooms = rooms.filter(
-				(room) => room.properties.Gebaeude === (building as Gebaeude)
-			)
-			if (buildingRooms.length === 0) {
-				return null
-			}
-			const floorCount = Array.from(
-				new Set(buildingRooms.map((room) => room.properties.Ebene))
-			).length
-			const location = buildingRooms[0].properties.Standort
-			const center = getCenter(buildingRooms.map((x) => x.geometry.coordinates))
-			return {
-				type: 'Feature',
-				id: building,
-				properties: {
-					Raum: building,
-					Funktion_en: 'Building',
-					Funktion_de: 'Gebäude',
-					Gebaeude: Gebaeude[building as keyof typeof Gebaeude],
-					Ebene: 'EG', // Dummy value to not break the floor picker
-					Etage: floorCount.toString(),
-					Standort: location,
-					rtype: SEARCH_TYPES.BUILDING,
-					center,
-					icon: getIcon(SEARCH_TYPES.BUILDING)
-				},
-				geometry: {
-					type: 'Point' as const,
-					coordinates: center
-				}
-			} satisfies Feature
-		}).filter(
-			(building): building is NonNullable<typeof building> => building !== null
-		)
-		return {
-			type: 'FeatureCollection',
-			features: [...rooms, ...buildings]
-		}
-	}, [mapOverlay])
-
-	const buildingGeoJSON: FeatureCollection = useMemo(() => {
-		return {
-			type: 'FeatureCollection',
-			features: allRooms.features.filter(
-				(f) => f.properties?.rtype === SEARCH_TYPES.BUILDING
-			)
-		}
-	}, [allRooms])
 
 	useEffect(() => {
 		// @ts-expect-error wrong type
@@ -340,51 +178,6 @@ const MapScreen = (): React.JSX.Element => {
 	}, [tabBarPressed])
 
 	useEffect(() => {
-		if (
-			params.room == null ||
-			params.room === '' ||
-			params.room === undefined
-		) {
-			return
-		}
-		if (
-			allRooms == null ||
-			allRooms.features.length === 0 ||
-			mapLoadState !== LoadingState.LOADED
-		) {
-			return
-		}
-
-		const room = allRooms.features.find(
-			(x) => x.properties?.Raum === params.room
-		)?.properties
-
-		if (room == null) {
-			roomNotFoundToast(params.room, theme.colors.notification)
-			router.setParams({ room: '' })
-			return
-		}
-		bottomSheetRef.current?.close()
-		setClickedElement({
-			data: params.room,
-			type: SEARCH_TYPES.ROOM,
-			center: room.center as Position | undefined,
-			manual: false
-		})
-		trackEvent('Room', {
-			room: params.room,
-			origin: 'InAppLink'
-		})
-		setCurrentFloor({
-			floor: (room.Ebene as string) ?? 'EG',
-			manual: false
-		})
-		handlePresentModalPress()
-
-		router.setParams({ room: '' })
-	}, [params, mapLoadState, allRooms])
-
-	useEffect(() => {
 		setMapCenter(
 			userFaculty === 'Nachhaltige Infrastruktur'
 				? NEUBURG_CENTER
@@ -393,110 +186,10 @@ const MapScreen = (): React.JSX.Element => {
 	}, [userFaculty])
 
 	useEffect(() => {
-		if (localSearch.length === 1 && params.room != null) {
-			router.setParams(undefined)
-		}
-	}, [localSearch])
-
-	const { data: roomStatusData } = useQuery({
-		queryKey: ['freeRooms', formatISODate(currentDate)],
-		queryFn: async () => await API.getFreeRooms(currentDate),
-		staleTime: 1000 * 60 * 60, // 60 minutes
-		gcTime: 1000 * 60 * 60 * 24 * 4, // 4 days
-		retry(failureCount, error) {
-			if (error instanceof NoSessionError) {
-				return false
-			}
-			return failureCount < 2
-		}
-	})
-
-	useEffect(() => {
-		function load(): void {
-			if (roomStatusData == null) {
-				console.debug('No room status data')
-				return
-			}
-			try {
-				const dateObj = new Date()
-				const date = formatISODate(dateObj)
-				const time = formatISOTime(dateObj)
-				const rooms = filterRooms(roomStatusData, date, time)
-				setAvailableRooms(rooms)
-				const openings = getRoomOpenings(roomStatusData, dateObj)
-				setRoomOpenings(openings)
-			} catch (e) {
-				if (
-					e instanceof NoSessionError ||
-					e instanceof UnavailableSessionError
-				) {
-					setAvailableRooms([])
-				} else {
-					console.error(e)
-				}
-			}
-		}
-		setAvailableRooms(null)
-		setRoomOpenings(null)
-		load()
-	}, [userKind, roomStatusData])
-
-	useEffect(() => {
 		if (clickedElement != null && currentFloor?.manual === true) {
 			bottomSheetModalRef.current?.close()
 		}
 	}, [currentFloor])
-
-	const uniqueEtages = Array.from(
-		new Set(
-			(Array.isArray(allRooms.features)
-				? allRooms.features
-				: Object.values(allRooms.features)
-			)
-				.map((room) => {
-					const properties = (room as RoomData).properties
-					return typeof properties?.Ebene === 'string' ? properties.Ebene : ''
-				})
-				.filter((etage) => etage != null)
-		)
-	).sort(
-		(a: string, b: string) => FLOOR_ORDER.indexOf(a) - FLOOR_ORDER.indexOf(b)
-	)
-	const [filteredGeoJSON, setFilteredGeoJSON] = useState<FeatureCollection>()
-	const [availableFilteredGeoJSON, setAvailableFilteredGeoJSON] =
-		useState<FeatureCollection>()
-
-	useEffect(() => {
-		if (mapOverlay == null) {
-			return
-		}
-		// filter the filteredGeoJSON to only show available rooms
-		const filteredFeatures = filterAvailableRooms(
-			filteredGeoJSON,
-			availableRooms
-		)
-
-		const availableFilteredGeoJSON: FeatureCollection = {
-			type: 'FeatureCollection',
-			features: filteredFeatures
-		}
-
-		setAvailableFilteredGeoJSON(availableFilteredGeoJSON)
-	}, [availableRooms, filteredGeoJSON, mapOverlay])
-
-	useEffect(() => {
-		if (mapOverlay == null) {
-			return
-		}
-		const filteredFeatures = filterEtage(currentFloor?.floor ?? 'EG', allRooms)
-
-		const newGeoJSON: FeatureCollection = {
-			...mapOverlay,
-			features: filteredFeatures
-		}
-
-		setFilteredGeoJSON(newGeoJSON)
-	}, [currentFloor, allRooms, mapOverlay])
 
 	const roomData: RoomData = useMemo(() => {
 		switch (clickedElement?.type) {
@@ -686,17 +379,6 @@ const MapScreen = (): React.JSX.Element => {
 		}
 	}, [regionChange, isVisible, opacity])
 
-	const showFiltered = useCallback(() => {
-		return filteredGeoJSON != null && filteredGeoJSON.features.length > 0
-	}, [filteredGeoJSON])
-
-	const showAvailableFiltered = useCallback(() => {
-		return (
-			availableFilteredGeoJSON != null &&
-			availableFilteredGeoJSON.features.length > 0
-		)
-	}, [availableFilteredGeoJSON])
-
 	const handleMapLoad = useCallback(() => {
 		setMapLoadState(LoadingState.LOADED)
 	}, [])
@@ -747,22 +429,15 @@ const MapScreen = (): React.JSX.Element => {
 					return
 				}
 
-				setClickedElement({
-					data: feature.properties.Raum,
-					type: SEARCH_TYPES.ROOM,
+				selectRoom({
+					room: feature.properties.Raum,
 					center,
+					origin: 'MapClick',
 					manual: true
 				})
-
-				trackEvent('Room', {
-					room: feature.properties.Raum,
-					origin: 'MapClick'
-				})
-
-				handlePresentModalPress()
 			}
 		},
-		[filteredGeoJSON, handlePresentModalPress]
+		[filteredGeoJSON, selectRoom]
 	)
 
 	const mapContainerStyle = {
@@ -802,7 +477,7 @@ const MapScreen = (): React.JSX.Element => {
 				>
 					<NavigationControl position="top-right" />
 
-					{showFiltered() && (
+					{hasFilteredRooms && (
 						<Source
 							id="allRoomsSource"
 							type="geojson"
@@ -829,7 +504,7 @@ const MapScreen = (): React.JSX.Element => {
 						</Source>
 					)}
 
-					{showAvailableFiltered() && (
+					{hasAvailableFilteredRooms && (
 						<Source
 							id="availableRoomsSource"
 							type="geojson"
